Data
I redid the GSTScript because I realized we needed to restructure the
BadgeBank table. Go ahead and rerun that on your server so that it is
correct.

I also created some INSERT statements to fill our database with Dummy
Data. This will help us build everything and make sure that it is
working